Fix for dtype

+++ onnx/backend/test/case/node/trilu.py
@@ -52,7 +52,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(-1).astype(np.int32)
+        k = np.array(-1).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -75,7 +75,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(-7).astype(np.int32)
+        k = np.array(-7).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -98,7 +98,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(2).astype(np.int32)
+        k = np.array(2).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -121,7 +121,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(6).astype(np.int32)
+        k = np.array(6).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -172,7 +172,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(2, 3, 3)).astype(np.int64)
-        k = np.array(-1).astype(np.int32)
+        k = np.array(-1).astype(np.int64)
         # X:
         # [[[4, 6, 9],
         #   [7, 5, 4],
@@ -201,7 +201,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(3, 1, 5)).astype(np.int64)
-        k = np.array(1).astype(np.int32)
+        k = np.array(1).astype(np.int64)
         # X:
         # [[[1, 4, 9, 7, 1]],
         #
@@ -226,7 +226,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(0, 5)).astype(np.int64)
-        k = np.array(6).astype(np.int32)
+        k = np.array(6).astype(np.int64)
         # X:
         # []
         # expect result:
@@ -267,7 +267,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(-1).astype(np.int32)
+        k = np.array(-1).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -291,7 +291,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(-7).astype(np.int32)
+        k = np.array(-7).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -315,7 +315,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(2).astype(np.int32)
+        k = np.array(2).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -338,7 +338,7 @@ class Trilu(Base):
             upper=0,
         )
         x = np.random.randint(10, size=(4, 5)).astype(np.int64)
-        k = np.array(6).astype(np.int32)
+        k = np.array(6).astype(np.int64)
         # X:
         #  [[4, 7, 3, 7, 9],
         #   [1, 2, 8, 6, 9],
@@ -391,7 +391,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(2, 3, 3)).astype(np.int64)
-        k = np.array(-1).astype(np.int32)
+        k = np.array(-1).astype(np.int64)
         # X:
         # [[[0, 4, 3],
         #   [2, 0, 9],
@@ -446,7 +446,7 @@ class Trilu(Base):
         )
 
         x = np.random.randint(10, size=(3, 0, 5)).astype(np.int64)
-        k = np.array(6).astype(np.int32)
+        k = np.array(6).astype(np.int64)
         # X:
         # []
         # expect result:
